+++ immutable_tree.go
@@ -148,8 +148,28 @@ func (t *ImmutableTree) Has(key []byte) (bool, error) {
 	return t.root.has(t, key)
 }
 
+func (t *ImmutableTree) hashAlgo() HashAlgo {
+	if t.ndb == nil {
+		return HashSHA256
+	}
+	return t.ndb.opts.hashAlgo()
+}
+
+func (t *ImmutableTree) useBlake3() bool {
+	return t.hashAlgo() == HashBLAKE3
+}
+
+func (t *ImmutableTree) newLeaf(key, value []byte) *Node {
+	n := NewNode(key, value)
+	n.algo = t.hashAlgo()
+	return n
+}
+
 // Hash returns the root hash.
 func (t *ImmutableTree) Hash() []byte {
+	if t.root == nil {
+		return emptyDigest(t.hashAlgo())
+	}
 	return t.root.hashWithCount(t.version + 1)
 }